fixup: move away from lib/ in tests as well
