+++ packages/hadron-document/test/editor/int64.spec.ts
@@ -1,6 +1,6 @@
 import { Long } from 'bson';
 import { Element } from '../..';
-import { Int64Editor } from '../../lib/editor';
+import { Int64Editor } from '../../src/editor';
 import { expect } from 'chai';
 
 describe('Int64Editor', function () {
